Update README.md
Changed the title name (based on the input from linuxonedemoaas squad channel). Also added two sections in the content. First, created a pre-req section, that depicts how to install the ibm cloud cli, which also now contains the steps to create an API key. The second section illustrates different steps on how to configure the application within an HPVS instance, using the git cli. Also, it walks the user through the pre-reqs for getting to the config on HPVS steps, such as locating the public ip, installing the git cli in HPVS, and changing directories to finish the 'run the application' segment.